Added fog to project

+++ MapGuessr/gameScene.h
@@ -10,7 +10,7 @@ class gameScene {
 private:
 	// This sets the zero point as map coord.
 	glm::vec2 worldStartingPoint = glm::vec2(51.59530934936532, 4.778385088898252);
-	glm::uvec2 worldIndexing = glm::uvec2(-1, -1);
+	glm::vec2 worldIndexing = glm::vec2(-1, -1);
 
 	std::shared_ptr<gameTile> gameTiles[viewDistance][viewDistance];
 	std::vector<std::shared_ptr<gameObject>> gameObjects;
@@ -21,7 +21,7 @@ private:
 	float calculateDistance(glm::vec3 l1, glm::vec3 l2);
 public:
 	gameScene(std::shared_ptr<fpsCamera> c);
-	void changedWorldIndex(glm::uvec2 worldIndex);
+	void changedWorldIndex(glm::vec2 worldIndex);
 
 	void draw();
 	void update(float elapsedTime);
